% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Correct_the_basis.R
\name{correctBasis4genTrend}
\alias{correctBasis4genTrend}
\title{Correct with the breed constant the basis of the extracted genTrend log-info}
\usage{
correctBasis4genTrend(ps_input_tibble, ps_par_bdiff, ps_output_csvfile)
}
\arguments{
\item{ps_input_tibble}{resulting tibble from the function extract4genTrend}

\item{ps_par_default}{path to the parameter file}
}
\value{
tbl_result resulting tibble with information extracted from genTrend log-file
}
\description{
Correct with the breed constant the basis of the extracted genTrend log-info
}
