% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read_and_Extract.R
\name{extract4genTrend}
\alias{extract4genTrend}
\title{Reading genTrend log-file and extract information to a csv-file}
\usage{
extract4genTrend(ps_input_logfile, ps_par_default)
}
\arguments{
\item{ps_input_logfile}{path to the input file coming from genTrend.f90}

\item{ps_par_default}{path to the parameter file}
}
\value{
tbl_result resulting tibble with information extracted from genTrend log-file
}
\description{
Read a genTrend file and extract information to a csv-file
}
